+++ django_x509/signals.py
@@ -1,0 +1,4 @@
+from django.dispatch import Signal
+
+x509_objects_expiring = Signal()
+x509_objects_expired = Signal()